requirements.txt
Pillow
opencv-python
pytesseract
google-cloud-translate
google-auth-oauthlib
numpy
pynput
pyinstaller
PyJWT